Feature quick map (#110)
* limitation parameter
this allows for a faster output, every <limitation> (e.g. 1000) route (with all of its trips that are not duplicates) will be output.

* Search for stops + headsign in lines
Adding search function for stops. Stops have the uic code and stop_name.
Also: amending the routes to include headsign and different colors for better visual separation

* Extending limitations to stops and highlighting for lines (against overlaps)

---------

Co-authored-by: u239901 <david.rudi2@sbb.ch>